requirements.txt
flask
pandas
python-telegram-bot==20.7
requests
